0415 保存

+++ prectice/prectice-workspace/prectice-workspace/src/main/java/org/example/juc/Sleeper.java
@@ -5,7 +5,7 @@ import lombok.SneakyThrows;
 public class Sleeper {
 
     @SneakyThrows
-    public static void sleep(int seconds) {
+    public static void sleep(long seconds) {
         Thread.sleep(1000 * seconds);
     }
 }
